admin/backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
aiohttp==3.9.1
pydantic==2.5.0
structlog==23.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and security
authlib==1.3.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
itsdangerous==2.1.2

# Session and caching
redis==5.0.1

# Additional utilities
python-dotenv==1.0.0
